Do Things Repeatedly

do() provides a natural syntax for repetition tuned to assist with replication and resampling methods.

do(object, ...)

# S3 method for class 'numeric'
do(object, ...)

# Default S3 method
do(object, ...)

Do(n = 1L, cull = NULL, mode = "default", algorithm = 1, parallel = TRUE)

# S3 method for class 'repeater'
print(x, ...)

# S4 method for class 'repeater,ANY'
e1 * e2

Arguments

object

an object

...

additional arguments

n

number of times to repeat

cull

function for culling output of objects being repeated. If NULL, a default culling function is used. The default culling function is currently aware of objects of types lme, lm, htest, table, cointoss, and matrix.

mode

target mode for value returned

algorithm

a number used to select the algorithm used. Currently numbers below 1 use an older algorithm and numbers >=1 use a newer algorithm which is faster in some situations.

parallel

a logical indicating whether parallel computation should be attempted using the parallel package (if it is installed and loaded).

x

an object created by do.

e1

an object (in cases documented here, the result of running do)

e2

an object (in cases documented here, an expression to be repeated)

Value

do returns an object of class repeater which is only useful in the context of the operator *. See the examples.

Note

do is a thin wrapper around Do to avoid collision with dplyr::do() from the dplyr package.

Naming

The names used in the object returned from do() are inferred from the objects created in each replication. Roughly, this the strategy employed.

If the objects have names, those names are inherited, if possible.

If the objects do not have names, but do() is used with a simple function call, the name of that function is used. Example: do(3) * mean(~height, data = Galton) produces a data frame with a variable named mean.

In cases where names are not easily inferred and a single result is produced, it is named result.

To get different names, one can rename the objects as they are created, or rename the result returned from do(). Example of the former: do(3) * c(mean_height = mean(~height, data = resample(Galton))).

See also

Author

Daniel Kaplan (kaplan@macalaster.edu) and Randall Pruim (rpruim@calvin.edu)

Examples

do(3) * rnorm(1)
#>        rnorm
#> 1 -0.7045291
#> 2  0.6722831
#> 3 -1.2826290
do(3) * "hello"
#>   hello
#> 1 hello
#> 2 hello
#> 3 hello
do(3) * 1:4
#>   V1 V2 V3 V4
#> 1  1  2  3  4
#> 2  1  2  3  4
#> 3  1  2  3  4
do(3) * mean(rnorm(25))
#>        mean
#> 1 0.1469577
#> 2 0.1167505
#> 3 0.1334250
do(3) * lm(shuffle(height) ~ sex + mother, Galton)
#>   Intercept       sexM      mother    sigma    r.squared         F numdf dendf
#> 1  69.36793 0.01015402 -0.04076656 3.585675 0.0006936883 0.3106410     2   895
#> 2  68.24441 0.13106505 -0.02421169 3.585847 0.0005978337 0.2676906     2   895
#> 3  64.50780 0.20611057  0.03348964 3.584681 0.0012479970 0.5591765     2   895
#>   .row .index
#> 1    1      1
#> 2    1      2
#> 3    1      3
do(3) * anova(lm(shuffle(height) ~ sex + mother, Galton))
#>                  source  df           SS        MS         F      pval .row
#> sex...1             sex   1    32.893416 32.893416 2.5669585 0.1094702    1
#> mother...2       mother   1    13.496052 13.496052 1.0532140 0.3050453    2
#> Residuals...3 Residuals 895 11468.672904 12.814160        NA        NA    3
#> sex...4             sex   1     8.665588  8.665588 0.6742219 0.4118022    1
#> mother...5       mother   1     3.208958  3.208958 0.2496714 0.6174293    2
#> Residuals...6 Residuals 895 11503.187826 12.852724        NA        NA    3
#> sex...7             sex   1     1.896518  1.896518 0.1476001 0.7009307    1
#> mother...8       mother   1    13.283584 13.283584 1.0338200 0.3095379    2
#> Residuals...9 Residuals 895 11499.882270 12.849030        NA        NA    3
#>               .index
#> sex...1            1
#> mother...2         1
#> Residuals...3      1
#> sex...4            2
#> mother...5         2
#> Residuals...6      2
#> sex...7            3
#> mother...8         3
#> Residuals...9      3
do(3) * c(sample.mean = mean(rnorm(25)))
#>   sample.mean
#> 1  0.13363120
#> 2  0.05743587
#> 3  0.19931276
# change the names on the fly
do(3) * mean(~height, data = resample(Galton))
#>       mean
#> 1 66.70234
#> 2 66.72038
#> 3 66.81114
do(3) * c(mean_height = mean(~height, data = resample(Galton)))
#>   mean_height
#> 1    66.88853
#> 2    66.38274
#> 3    66.75523
set.rseed(1234)
do(3) * tally( ~sex|treat, data=resample(HELPrct))
#>   female.no male.no female.yes male.yes
#> 1        45     184         59      165
#> 2        52     179         58      164
#> 3        52     163         43      195
set.rseed(1234)  # re-using seed gives same results again
do(3) * tally( ~sex|treat, data=resample(HELPrct))
#>   female.no male.no female.yes male.yes
#> 1        45     184         59      165
#> 2        52     179         58      164
#> 3        52     163         43      195