Check if formula

Check if formula

Check for simple formula

Extract simple part from formula

.is.formula(x)

.is.simple.formula(x)

.simple.part(x)

.make.data.frame(x)

.merge_data_frames(a, b)

.squash_names(object, sep = ":")

.do.safe.call(what, args, quote = FALSE, envir = parent.frame(), ...)

Arguments

x

object to be converted

a

a data frame

b

a data frame

object

an object

sep

a character

what

either a function or a non-empty character string naming the function to be called.

args

a list of arguments to the function call. The names attribute of args gives the argument names.

quote

a logical value indicating whether to quote the arguments.

envir

an environment within which to evaluate the call. This will be most useful if what is a character string and the arguments are symbols or quoted expressions.

Value

TRUE for a formula, FALSE otherwise, even if evaluation throws an error

TRUE if formula has no left-hand side or a simple right-hand side (e.g., NULL, ., 1, or 0)

simple part of formula or NULL if formula is not simple

a data frame

a data frame

a character vector

The result of the (evaluated) function call.

Details

.make.data.frame converts things to a data frame

.merge_data_frames is a wrapper around merge

.squash_names squashes names of a data frame into a single string

.do.safe.call avoids conflicts between named arguments and ... by taking named arguments preferentially.

See also