Mosaic Server

Creates and registers a global mosaic API, coordinator and connector that can be acessed from js using:

Usage

mosaicServer(id, connection)

Arguments

id

The id of the module.

connection

A DBI connection to a DuckDB database.

Value

A shiny reactive that returns the id of the API.

Details

window.[id]
window.[id + "_coordinator"]
window.[id + "_connector"]

This function is intended to be used in Shiny apps in combination with the api parameter of the mosaic() function. Allowing you to create plots that fetch data from a server-side DuckDB database, so you don't need to send large datasets to the client.