The Data Distribution

Density, distribution function, quantile function, and random generation from data.

qdata(formula, p = seq(0, 1, 0.25), data = NULL, ...)

cdata(formula, p = 0.95, data = NULL, ...)

pdata(formula, q, data = NULL, ...)

rdata(formula, n, data = NULL, ...)

ddata(formula, q, data = NULL, ...)

Arguments

formula
a formula or a vector

p
a vector of probabilities

data
a data frame in which to evaluate formula

…
additional arguments passed to quantile or sample

q
a vector of quantiles

n
number of values to sample

Value

For qdata, a vector of quantiles

for cdata, a named numerical vector or a data frame giving upper and lower limits and the central proportion requested

For pdata, a vector of probabilities

For rdata, a vector of sampled values.

For ddata, a vector of probabilities (empirical densities)

Examples

data(iris)
qdata(Sepal.Length ~ Species, 0.5, data=iris)

#>      Species   p quantile
#> 1     setosa 0.5      5.0
#> 2 versicolor 0.5      5.9
#> 3  virginica 0.5      6.5

qdata(~Sepal.Length, p = 0.5, groups=Species, data=iris)

#>      Species   p quantile
#> 1     setosa 0.5      5.0
#> 2 versicolor 0.5      5.9
#> 3  virginica 0.5      6.5

qdata(iris$Sepal.Length, p = 0.5)

#>        p quantile 
#>      0.5      5.8 

qdata(~ Sepal.Length, p = 0.5, data=iris)

#>        p quantile 
#>      0.5      5.8 

qdata(~ Sepal.Length, p = 0.5, groups=Species, data=iris)

#>      Species   p quantile
#> 1     setosa 0.5      5.0
#> 2 versicolor 0.5      5.9
#> 3  virginica 0.5      6.5

data(iris)
cdata(iris$Sepal.Length, 0.5)

#>       low        hi central.p 
#>       5.1       6.4       0.5 

cdata( ~ Sepal.Length, 0.5, data = iris)

#>       low        hi central.p 
#>       5.1       6.4       0.5 

cdata( ~ Sepal.Length, 0.5, data = iris)

#>       low        hi central.p 
#>       5.1       6.4       0.5 

cdata( ~ Sepal.Length | Species, data = iris, p = .5)

#>      Species   low  hi central.p
#> 1     setosa 4.800 5.2       0.5
#> 2 versicolor 5.600 6.3       0.5
#> 3  virginica 6.225 6.9       0.5

data(iris)
pdata(iris$Sepal.Length, 3:6)

#> [1] 0.0000000 0.0000000 0.2133333 0.5933333

pdata( ~ Sepal.Length, 3:6, data=iris)

#> [1] 0.0000000 0.0000000 0.2133333 0.5933333

data(iris)
rdata(iris$Species, 10)

#>  [1] versicolor versicolor setosa     setosa     setosa     setosa    
#>  [7] setosa     versicolor versicolor versicolor
#> Levels: setosa versicolor virginica

rdata(~Species, n = 10, data=iris)

#>  [1] virginica  setosa     setosa     virginica  virginica  virginica 
#>  [7] setosa     versicolor setosa     setosa    
#> Levels: setosa versicolor virginica

rdata(Sepal.Length ~ Species,  n = 5, data=iris)

#>      Species rdata_v1 rdata_v2 rdata_v3 rdata_v4 rdata_v5
#> 1     setosa      5.4      5.0      5.4      5.4      5.0
#> 2 versicolor      6.0      5.0      5.5      6.0      5.6
#> 3  virginica      5.7      6.7      6.5      7.3      6.7

data(iris)
ddata(iris$Species, 'setosa')

#>    setosa 
#> 0.3333333 

ddata(~Species, 'setosa', data=iris)

#>    setosa 
#> 0.3333333