Do Things Repeatedly

do() provides a natural syntax for repetition tuned to assist with replication and resampling methods.

do(object, ...)

# S3 method for numeric
do(object, ...)

# S3 method for default
do(object, ...)

Do(n = 1L, cull = NULL, mode = "default", algorithm = 1, parallel = TRUE)

# S3 method for repeater
print(x, ...)

# S4 method for repeater,ANY
*(e1, e2)

Arguments

object
an object

...
additional arguments

n
number of times to repeat

cull
function for culling output of objects being repeated. If NULL, a default culling function is used. The default culling function is currently aware of objects of types lme, lm, htest, table, cointoss, and matrix.

mode
target mode for value returned

algorithm
a number used to select the algorithm used. Currently numbers below 1 use an older algorithm and numbers >=1 use a newer algorithm which is faster in some situations.

parallel
a logical indicating whether parallel computation should be attempted using the parallel package (if it is installed and loaded).

x
an object created by do.

e1
an object (in cases documented here, the result of running do)

e2
an object (in cases documented here, an expression to be repeated)

Value

do returns an object of class repeater which is only useful in the context of the operator *. See the examples.

Note

do is a thin wrapper around Do to avoid collision with dplyr::do() from the dplyr package.

Naming

The names used in the object returned from do() are inferred from the objects created in each replication. Roughly, this the strategy employed.

If the objects have names, those names are inherited, if possible.

If the objects do not have names, but do() is used with a simple function call, the name of that function is used. Example: do(3) * mean(~height, data = Galton) produces a data frame with a variable named mean.

In cases where names are not easily inferred and a single result is produced, it is named result.

To get different names, one can rename the objects as they are created, or rename the result returned from do(). Example of the former: do(3) * c(mean_height = mean(~height, data = resample(Galton))).

See also

Author

Daniel Kaplan (kaplan@macalaster.edu) and Randall Pruim (rpruim@calvin.edu)

Examples

do(3) * rnorm(1)

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>        rnorm
#> 1 -1.9633309
#> 2  0.6799135
#> 3 -0.9810619

do(3) * "hello"

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   hello
#> 1 hello
#> 2 hello
#> 3 hello

do(3) * 1:4

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   V1 V2 V3 V4
#> 1  1  2  3  4
#> 2  1  2  3  4
#> 3  1  2  3  4

do(3) * mean(rnorm(25))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>         mean
#> 1  0.4051522
#> 2  0.2165816
#> 3 -0.1151129

do(3) * lm(shuffle(height) ~ sex + mother, Galton)

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   Intercept        sexM       mother    sigma    r.squared        F numdf dendf
#> 1  68.63566 -0.07571604 -0.028645974 3.586134 0.0004379646 0.196075     2   895
#> 2  69.37194  0.08719953 -0.041451671 3.585334 0.0008835761 0.395750     2   895
#> 3  66.54729  0.14832865  0.002131402 3.586151 0.0004282589 0.191728     2   895
#>   .row .index
#> 1    1      1
#> 2    1      2
#> 3    1      3

do(3) * anova(lm(shuffle(height) ~ sex + mother, Galton))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>                  source  df           SS        MS         F      pval .row
#> sex...1             sex   1     1.878168  1.878168 0.1460253 0.7024538    1
#> mother...2       mother   1     1.747952  1.747952 0.1359011 0.7124775    2
#> Residuals...3 Residuals 895 11511.436251 12.861940        NA        NA    3
#> sex...4             sex   1     6.501050  6.501050 0.5061294 0.4770046    1
#> mother...5       mother   1    12.608856 12.608856 0.9816434 0.3220612    2
#> Residuals...6 Residuals 895 11495.952466 12.844640        NA        NA    3
#> sex...7             sex   1    17.266731 17.266731 1.3464580 0.2462077    1
#> mother...8       mother   1    20.479387 20.479387 1.5969806 0.2066607    2
#> Residuals...9 Residuals 895 11477.316254 12.823817        NA        NA    3
#>               .index
#> sex...1            1
#> mother...2         1
#> Residuals...3      1
#> sex...4            2
#> mother...5         2
#> Residuals...6      2
#> sex...7            3
#> mother...8         3
#> Residuals...9      3

do(3) * c(sample.mean = mean(rnorm(25)))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   sample.mean
#> 1 -0.47640160
#> 2  0.07563422
#> 3 -0.08604106

# change the names on the fly
do(3) * mean(~height, data = resample(Galton))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>       mean
#> 1 66.73296
#> 2 66.66047
#> 3 66.74777

do(3) * c(mean_height = mean(~height, data = resample(Galton)))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   mean_height
#> 1    66.77906
#> 2    66.74521
#> 3    66.73274

set.rseed(1234)
do(3) * tally( ~sex|treat, data=resample(HELPrct))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   female.no male.no female.yes male.yes
#> 1        56     181         49      167
#> 2        53     162         58      180
#> 3        53     188         43      169

set.rseed(1234)  # re-using seed gives same results again
do(3) * tally( ~sex|treat, data=resample(HELPrct))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   female.no male.no female.yes male.yes
#> 1        56     181         49      167
#> 2        53     162         58      180
#> 3        53     188         43      169