Do Things Repeatedly

do() provides a natural syntax for repetition tuned to assist with replication and resampling methods.

do(object, ...)

# S3 method for numeric
do(object, ...)

# S3 method for default
do(object, ...)

Do(n = 1L, cull = NULL, mode = "default", algorithm = 1, parallel = TRUE)

# S3 method for repeater
print(x, ...)

# S4 method for repeater,ANY
*(e1, e2)

Arguments

object
an object

...
additional arguments

n
number of times to repeat

cull
function for culling output of objects being repeated. If NULL, a default culling function is used. The default culling function is currently aware of objects of types lme, lm, htest, table, cointoss, and matrix.

mode
target mode for value returned

algorithm
a number used to select the algorithm used. Currently numbers below 1 use an older algorithm and numbers >=1 use a newer algorithm which is faster in some situations.

parallel
a logical indicating whether parallel computation should be attempted using the parallel package (if it is installed and loaded).

x
an object created by do.

e1
an object (in cases documented here, the result of running do)

e2
an object (in cases documented here, an expression to be repeated)

Value

do returns an object of class repeater which is only useful in the context of the operator *. See the examples.

Note

do is a thin wrapper around Do to avoid collision with dplyr::do() from the dplyr package.

Naming

The names used in the object returned from do() are inferred from the objects created in each replication. Roughly, this the strategy employed.

If the objects have names, those names are inherited, if possible.

If the objects do not have names, but do() is used with a simple function call, the name of that function is used. Example: do(3) * mean(~height, data = Galton) produces a data frame with a variable named mean.

In cases where names are not easily inferred and a single result is produced, it is named result.

To get different names, one can rename the objects as they are created, or rename the result returned from do(). Example of the former: do(3) * c(mean_height = mean(~height, data = resample(Galton))).

See also

Author

Daniel Kaplan (kaplan@macalaster.edu) and Randall Pruim (rpruim@calvin.edu)

Examples

do(3) * rnorm(1)

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>        rnorm
#> 1 -1.0345478
#> 2  0.7987134
#> 3  0.1053965

do(3) * "hello"

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   hello
#> 1 hello
#> 2 hello
#> 3 hello

do(3) * 1:4

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   V1 V2 V3 V4
#> 1  1  2  3  4
#> 2  1  2  3  4
#> 3  1  2  3  4

do(3) * mean(rnorm(25))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>           mean
#> 1  0.006721584
#> 2  0.039348176
#> 3 -0.192240406

do(3) * lm(shuffle(height) ~ sex + mother, Galton)

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   Intercept       sexM     mother    sigma   r.squared         F numdf dendf
#> 1  64.45391 -0.1710129 0.03737782 3.584785 0.001189603 0.5329814     2   895
#> 2  64.45215 -0.3852502 0.03913629 3.580419 0.003621433 1.6264814     2   895
#> 3  64.21043 -0.1280755 0.04083021 3.585047 0.001044051 0.4677011     2   895
#>   .row .index
#> 1    1      1
#> 2    1      2
#> 3    1      3

do(3) * anova(lm(shuffle(height) ~ sex + mother, Galton))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>                  source  df           SS         MS          F      pval .row
#> sex...1             sex   1 5.919968e-01  0.5919968 0.04610984 0.8300252    1
#> mother...2       mother   1 2.371030e+01 23.7103009 1.84676376 0.1745026    2
#> Residuals...3 Residuals 895 1.149076e+04 12.8388381         NA        NA    3
#> sex...4             sex   1 1.842778e+01 18.4277826 1.43510543 0.2312502    1
#> mother...5       mother   1 4.192925e+00  4.1929247 0.32653353 0.5678509    2
#> Residuals...6 Residuals 895 1.149244e+04 12.8407169         NA        NA    3
#> sex...7             sex   1 4.488690e+01 44.8868984 3.50742895 0.0614200    1
#> mother...8       mother   1 1.626590e+01 16.2659019 1.27100551 0.2598803    2
#> Residuals...9 Residuals 895 1.145391e+04 12.7976643         NA        NA    3
#>               .index
#> sex...1            1
#> mother...2         1
#> Residuals...3      1
#> sex...4            2
#> mother...5         2
#> Residuals...6      2
#> sex...7            3
#> mother...8         3
#> Residuals...9      3

do(3) * c(sample.mean = mean(rnorm(25)))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   sample.mean
#> 1  0.08799766
#> 2  0.14830201
#> 3  0.09310375

# change the names on the fly
do(3) * mean(~height, data = resample(Galton))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>       mean
#> 1 66.63051
#> 2 66.67105
#> 3 66.68964

do(3) * c(mean_height = mean(~height, data = resample(Galton)))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   mean_height
#> 1    66.71125
#> 2    66.65445
#> 3    66.92584

set.rseed(1234)
do(3) * tally( ~sex|treat, data=resample(HELPrct))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   female.no male.no female.yes male.yes
#> 1        56     181         49      167
#> 2        53     162         58      180
#> 3        53     188         43      169

set.rseed(1234)  # re-using seed gives same results again
do(3) * tally( ~sex|treat, data=resample(HELPrct))

#> Using parallel package.
#>   * Set seed with set.rseed().
#>   * Disable this message with options(`mosaic:parallelMessage` = FALSE)

#>   female.no male.no female.yes male.yes
#> 1        56     181         49      167
#> 2        53     162         58      180
#> 3        53     188         43      169