Do Things Repeatedly

do() provides a natural syntax for repetition tuned to assist with replication and resampling methods.

do(object, ...)

# S3 method for numeric
do(object, ...)

# S3 method for default
do(object, ...)

Do(n = 1L, cull = NULL, mode = "default", algorithm = 1,
  parallel = TRUE)

# S3 method for repeater
print(x, ...)

# S4 method for repeater,ANY
*(e1, e2)

Arguments

object
an object

...
additional arguments

n
number of times to repeat

cull
function for culling output of objects being repeated. If NULL, a default culling function is used. The default culling function is currently aware of objects of types lme, lm, htest, table, cointoss, and matrix.

mode
target mode for value returned

algorithm
a number used to select the algorithm used. Currently numbers below 1 use an older algorithm and numbers >=1 use a newer algorithm which is faster in some situations.

parallel
a logical indicating whether parallel computation should be attempted using the parallel package (if it is installed and loaded).

x
an object created by do.

e1
an object (in cases documented here, the result of running do)

e2
an object (in cases documented here, an expression to be repeated)

Value

do returns an object of class repeater which is only useful in the context of the operator *. See the examples.

Note

do is a thin wrapper around Do to avoid collision with dplyr::do() from the dplyr package.

Naming

The names used in the object returned from do() are inferred from the objects created in each replication. Roughly, this the strategy employed.

If the objects have names, those names are inherited, if possible.

If the objects do not have names, but do() is used with a simple function call, the name of that function is used. Example: do(3) * mean(~height, data = Galton) produces a data frame with a variable named mean.

In cases where names are not easily inferred and a single result is produced, it is named result.

To get different names, one can rename the objects as they are created, or rename the result returned from do(). Example of the former: do(3) * c(mean_height = mean(~height, data = resample(Galton))).

See also

Examples

do(3) * rnorm(1)

#>        rnorm
#> 1  0.1237653
#> 2  0.9667625
#> 3 -0.6537146

do(3) * "hello"

#>   hello
#> 1 hello
#> 2 hello
#> 3 hello

do(3) * 1:4

#>   V1 V2 V3 V4
#> 1  1  2  3  4
#> 2  1  2  3  4
#> 3  1  2  3  4

do(3) * mean(rnorm(25))

#>          mean
#> 1  0.38409739
#> 2 -0.19761655
#> 3 -0.03594621

do(3) * lm(shuffle(height) ~ sex + mother, Galton)

#>   Intercept        sexM       mother    sigma    r.squared          F numdf
#> 1  66.78553  0.08179791 -0.001048568 3.586684 0.0001312863 0.05875833     2
#> 2  73.36679 -0.19384099 -0.101518027 3.578159 0.0048785224 2.19384147     2
#> 3  65.54927  0.36245417  0.015974805 3.582205 0.0026267676 1.17857434     2
#>   dendf .row .index
#> 1   895    1      1
#> 2   895    1      2
#> 3   895    1      3

do(3) * anova(lm(shuffle(height) ~ sex + mother, Galton))

#>      source  df           SS         MS           F      pval .row .index
#> 1       sex   1 4.742749e+01 47.4274869 3.709896389 0.0544072    1      1
#> 2    mother   1 2.591351e+01 25.9135094 2.027019378 0.1548719    2      1
#> 3 Residuals 895 1.144172e+04 12.7840462          NA        NA    3      1
#> 4       sex   1 4.035162e+00  4.0351615 0.313740651 0.5755340    1      2
#> 5    mother   1 2.422190e-02  0.0242219 0.001883294 0.9653948    2      2
#> 6 Residuals 895 1.151100e+04 12.8614559          NA        NA    3      2
#> 7       sex   1 1.163642e+01 11.6364221 0.905424974 0.3415881    1      3
#> 8    mother   1 9.845130e-01  0.9845130 0.076604533 0.7820166    2      3
#> 9 Residuals 895 1.150244e+04 12.8518899          NA        NA    3      3

do(3) * c(sample.mean = mean(rnorm(25)))

#>   sample.mean
#> 1  0.05592448
#> 2  0.16850181
#> 3  0.18911600

# change the names on the fly
do(3) * mean(~height, data = resample(Galton))

#>       mean
#> 1 66.88318
#> 2 66.60947
#> 3 66.79610

do(3) * c(mean_height = mean(~height, data = resample(Galton)))

#>   mean_height
#> 1    66.70612
#> 2    66.63018
#> 3    66.77205

set.rseed(1234)
do(3) * tally( ~sex|treat, data=resample(HELPrct))

#>   female.no male.no female.yes male.yes
#> 1        47     168         49      189
#> 2        58     157         43      195
#> 3        52     179         55      167

set.rseed(1234)  # re-using seed gives same results again
do(3) * tally( ~sex|treat, data=resample(HELPrct))

#>   female.no male.no female.yes male.yes
#> 1        47     168         49      189
#> 2        58     157         43      195
#> 3        52     179         55      167