The Data Distribution

Utility functions for density, distribution function, quantile function, and random generation from data.

qdata_v(x, p = seq(0, 1, 0.25), na.rm = TRUE, ...)

qdata_f(x, ..., data = NULL, groups = NULL, na.rm = TRUE)

cdata_v(x, p = 0.95, na.rm = TRUE, ...)

cdata_f(x, ..., data = NULL, groups = NULL, na.rm = TRUE)

pdata_v(x, q, lower.tail = TRUE, ...)

pdata_f(x, ..., data = NULL, groups = NULL, na.rm = TRUE)

rdata_v(x, n, replace = TRUE, ...)

rdata_f(x, ..., data = NULL, groups = NULL, na.rm = TRUE)

ddata_v(x, q, ..., data = NULL, log = FALSE, na.rm = TRUE)

ddata_f(x, ..., data = NULL, groups = NULL, na.rm = TRUE)

Arguments

x
a vector containing the data

p
a vector of probabilities

na.rm
a logical indicating whether NAs should be removed before computing.

...
additional arguments passed to quantile or sample

data
a data frame in which to evaluate formula

groups
a grouping variable, typically the name of a variable in data

q
a vector of quantiles

lower.tail
a logical indicating whether to use the lower or upper tail probability

n
number of values to sample

replace
a logical indicating whether to sample with replacement

log
a logical indicating whether the result should be log transformed

See also