More Random Samples

These functions simplify and unify sampling in various ways.

resample(..., replace = TRUE)

deal(...)

shuffle(x, replace = FALSE, prob = NULL, groups = NULL, orig.ids = FALSE)

sample(x, size, replace = FALSE, ...)

# S3 method for default
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = FALSE,
  ...
)

# S3 method for data.frame
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = TRUE,
  fixed = names(x),
  shuffled = c(),
  invisibly.return = NULL,
  ...
)

# S3 method for matrix
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = FALSE,
  ...
)

# S3 method for factor
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = FALSE,
  drop.unused.levels = FALSE,
  ...
)

Arguments

...
additional arguments passed to base::sample() or mosaic::sample().

replace
Should sampling be with replacement?

x
Either a vector of one or more elements from which to choose, or a positive integer.

prob
A vector of probability weights for obtaining the elements of the vector being sampled.

groups
a vector (or variable in a data frame) specifying groups to sample within. This will be recycled if necessary.

orig.ids
a logical; should original ids be included in returned data frame?

size
a non-negative integer giving the number of items to choose.

fixed
a vector of column names. These variables are shuffled en masse, preserving associations among these columns.

shuffled
a vector of column names. these variables are reshuffled individually (within groups if groups is specified), breaking associations among these columns. examples.

invisibly.return
a logical, should return be invisible?

drop.unused.levels
a logical, should unused levels be dropped?

Details

These functions are wrappers around sample() providing different defaults and natural names.

Examples

# 100 Bernoulli trials -- no need for replace=TRUE
resample(0:1, 100)

#>   [1] 1 1 1 1 1 1 1 0 0 0 1 0 0 1 1 0 0 1 0 0 0 1 1 1 1 0 1 1 1 0 1 1 0 1 1 1 0
#>  [38] 1 1 0 0 1 0 0 1 1 0 0 0 1 1 1 0 0 1 1 0 0 1 1 1 0 0 0 1 0 1 1 1 0 0 0 1 1
#>  [75] 1 1 0 1 0 1 0 1 1 1 0 0 0 0 0 0 0 0 0 1 0 1 0 1 1 0

tally(resample(0:1, 100))

#> Warning: `data_frame()` is deprecated as of tibble 1.1.0.
#> Please use `tibble()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> X
#>  0  1 
#> 54 46 

if (require(mosaicData)) {
Small <- sample(KidsFeet, 10)
resample(Small)
tally(~ sex, data=resample(Small))
tally(~ sex, data=resample(Small))
# fixed marginals for sex
tally(~ sex, data=Small)
tally(~ sex, data=resample(Small, groups=sex))
# shuffled can be used to reshuffle some variables within groups
# orig.id shows where the values were in original data frame.
Small <- mutate(Small,
   id1 = paste(sex,1:10, sep=":"),
   id2 = paste(sex,1:10, sep=":"))
resample(Small, groups=sex, shuffled=c("id1","id2"))
}

#>          name birthmonth birthyear length width sex biggerfoot domhand orig.id
#> 1       David          5        88   24.4   8.4   B          L       R  10.3.3
#> 2        Lars         10        87   25.4   8.8   B          L       L   8.8.8
#> 22       Josh          7        88   24.4   8.6   B          L       R   3.8.8
#> 2.1      Lars         10        87   25.4   8.8   B          L       L  8.10.4
#> 13        Cal          8        87   26.1   9.1   B          L       R  4.8.10
#> 22.1     Josh          7        88   24.4   8.6   B          L       R   3.3.8
#> 2.2      Lars         10        87   25.4   8.8   B          L       L   8.4.3
#> 31   Danielle          6        88   24.0   9.3   G          L       R   5.5.5
#> 31.1 Danielle          6        88   24.0   9.3   G          L       R   5.5.6
#> 32    Caitlin          7        88   22.5   8.6   G          R       R   6.6.5
#>       id1  id2
#> 1     B:3  B:3
#> 2     B:8  B:8
#> 22    B:8  B:8
#> 2.1  B:10  B:4
#> 13    B:8 B:10
#> 22.1  B:3  B:8
#> 2.2   B:4  B:3
#> 31    G:5  G:5
#> 31.1  G:5  G:6
#> 32    G:6  G:5

deal(Cards, 13)    # A Bridge hand

#>  [1] "5H" "AD" "4S" "6H" "5D" "JD" "AH" "7D" "4C" "9H" "QD" "5S" "6S"

shuffle(Cards)

#>  [1] "6H"  "QC"  "QS"  "4C"  "KC"  "AH"  "KD"  "10C" "2D"  "KS"  "3C"  "4H" 
#> [13] "9S"  "QD"  "10D" "QH"  "9C"  "JH"  "8C"  "8D"  "2C"  "5S"  "7H"  "9H" 
#> [25] "6D"  "KH"  "3S"  "9D"  "8H"  "7S"  "2H"  "7C"  "AC"  "JD"  "4D"  "3H" 
#> [37] "6S"  "5H"  "10H" "3D"  "JC"  "5D"  "10S" "JS"  "5C"  "AS"  "6C"  "7D" 
#> [49] "4S"  "AD"  "8S"  "2S"