More Random Samples

These functions simplify and unify sampling in various ways.

resample(..., replace = TRUE)

deal(...)

shuffle(x, replace = FALSE, prob = NULL, groups = NULL, orig.ids = FALSE)

sample(x, size, replace = FALSE, ...)

# S3 method for default
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = FALSE,
  ...
)

# S3 method for data.frame
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = TRUE,
  fixed = names(x),
  shuffled = c(),
  invisibly.return = NULL,
  ...
)

# S3 method for matrix
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = FALSE,
  ...
)

# S3 method for factor
sample(
  x,
  size,
  replace = FALSE,
  prob = NULL,
  groups = NULL,
  orig.ids = FALSE,
  drop.unused.levels = FALSE,
  ...
)

Arguments

...
additional arguments passed to base::sample() or mosaic::sample().

replace
Should sampling be with replacement?

x
Either a vector of one or more elements from which to choose, or a positive integer.

prob
A vector of probability weights for obtaining the elements of the vector being sampled.

groups
a vector (or variable in a data frame) specifying groups to sample within. This will be recycled if necessary.

orig.ids
a logical; should original ids be included in returned data frame?

size
a non-negative integer giving the number of items to choose.

fixed
a vector of column names. These variables are shuffled en masse, preserving associations among these columns.

shuffled
a vector of column names. these variables are reshuffled individually (within groups if groups is specified), breaking associations among these columns. examples.

invisibly.return
a logical, should return be invisible?

drop.unused.levels
a logical, should unused levels be dropped?

Details

These functions are wrappers around sample() providing different defaults and natural names.

Examples

# 100 Bernoulli trials -- no need for replace=TRUE
resample(0:1, 100)

#>   [1] 1 1 0 0 1 0 0 1 0 1 1 0 0 1 0 1 1 1 0 1 0 1 1 0 0 1 0 1 0 0 1 0 0 0 0 1 1
#>  [38] 1 1 1 1 0 0 0 0 1 0 1 0 1 1 1 1 1 1 1 0 0 1 1 1 1 1 0 1 1 1 1 1 1 0 0 0 1
#>  [75] 1 0 1 0 1 1 1 0 1 1 1 0 0 1 1 0 0 1 0 1 1 0 0 1 1 0

tally(resample(0:1, 100))

#> X
#>  0  1 
#> 53 47 

if (require(mosaicData)) {
Small <- sample(KidsFeet, 10)
resample(Small)
tally(~ sex, data=resample(Small))
tally(~ sex, data=resample(Small))
# fixed marginals for sex
tally(~ sex, data=Small)
tally(~ sex, data=resample(Small, groups=sex))
# shuffled can be used to reshuffle some variables within groups
# orig.id shows where the values were in original data frame.
Small <- mutate(Small,
   id1 = paste(sex,1:10, sep=":"),
   id2 = paste(sex,1:10, sep=":"))
resample(Small, groups=sex, shuffled=c("id1","id2"))
}

#>         name birthmonth birthyear length width sex biggerfoot domhand orig.id
#> 2     Scotty          3        88   25.7   9.7   B          R       R   2.2.2
#> 2.1   Scotty          3        88   25.7   9.7   B          R       R   2.2.2
#> 5      Julie         11        87   26.0   9.3   G          L       R   5.5.6
#> 5.1    Julie         11        87   26.0   9.3   G          L       R   5.6.4
#> 8      Laura          9        88   24.0   8.3   G          R       L   8.4.8
#> 4     Hayley          1        88   21.6   7.9   G          R       R   4.5.1
#> 7   Danielle          6        88   24.0   9.3   G          L       R   7.1.7
#> 1    Caitlin          6        88   23.0   8.8   G          L       R   1.5.5
#> 6       Abby          2        88   26.1   9.5   G          L       R   6.7.5
#> 5.2    Julie         11        87   26.0   9.3   G          L       R   5.8.5
#>     id1 id2
#> 2   B:2 B:2
#> 2.1 B:2 B:2
#> 5   G:5 G:6
#> 5.1 G:6 G:4
#> 8   G:4 G:8
#> 4   G:5 G:1
#> 7   G:1 G:7
#> 1   G:5 G:5
#> 6   G:7 G:5
#> 5.2 G:8 G:5

deal(Cards, 13)    # A Bridge hand

#>  [1] "AD"  "3S"  "KS"  "7C"  "2H"  "QD"  "6S"  "JH"  "10H" "AC"  "5S"  "KH" 
#> [13] "3H" 

shuffle(Cards)

#>  [1] "8H"  "9H"  "10H" "8C"  "AD"  "7C"  "4C"  "AC"  "6D"  "JC"  "6H"  "4H" 
#> [13] "4S"  "3C"  "7D"  "2D"  "QD"  "2H"  "10D" "JS"  "KH"  "6C"  "JD"  "4D" 
#> [25] "KS"  "QS"  "5C"  "9C"  "2C"  "5S"  "QC"  "10C" "3S"  "8S"  "KC"  "6S" 
#> [37] "9D"  "QH"  "9S"  "5H"  "3D"  "8D"  "KD"  "10S" "2S"  "7H"  "JH"  "AS" 
#> [49] "3H"  "5D"  "7S"  "AH"