More Random Samples

These functions simplify and unify sampling in various ways.

resample(..., replace = TRUE)

deal(...)

shuffle(x, replace = FALSE, prob = NULL, groups = NULL,
  orig.ids = FALSE)

sample(x, size, replace = FALSE, ...)

# S3 method for default
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = FALSE, ...)

# S3 method for data.frame
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = TRUE, fixed = names(x), shuffled = c(),
  invisibly.return = NULL, ...)

# S3 method for matrix
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = FALSE, ...)

# S3 method for factor
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = FALSE, drop.unused.levels = FALSE, ...)

Arguments

…
additional arguments passed to base::sample() or mosaic::sample().

replace
Should sampling be with replacement?

x
Either a vector of one or more elements from which to choose, or a positive integer.

prob
A vector of probability weights for obtaining the elements of the vector being sampled.

groups
a vector (or variable in a data frame) specifying groups to sample within. This will be recycled if necessary.

orig.ids
a logical; should original ids be included in returned data frame?

size
a non-negative integer giving the number of items to choose.

fixed
a vector of column names. These variables are shuffled en masse, preserving associations among these columns.

shuffled
a vector of column names. these variables are reshuffled individually (within groups if groups is specified), breaking associations among these columns. examples.

invisibly.return
a logical, should return be invisible?

drop.unused.levels
a logical, should unused levels be dropped?

Details

These functions are wrappers around sample() providing different defaults and natural names.

Examples

# 100 Bernoulli trials -- no need for replace=TRUE
resample(0:1, 100)

#>   [1] 1 1 1 1 1 0 0 1 0 0 0 0 1 0 1 1 0 0 0 0 0 1 0 0 0 0 1 1 0 0 1 1 0 0 1 0 0
#>  [38] 0 1 0 0 0 1 1 1 0 1 0 0 1 0 1 1 1 1 0 0 1 1 1 1 1 1 0 1 1 1 1 0 0 1 0 1 0
#>  [75] 1 1 1 1 0 0 0 0 1 0 0 1 0 0 1 1 1 1 1 1 0 1 1 1 1 1

tally(resample(0:1, 100))

#> X
#>  0  1 
#> 51 49 

if (require(mosaicData)) {
Small <- sample(KidsFeet, 10)
resample(Small)
tally(~ sex, data=resample(Small))
tally(~ sex, data=resample(Small))
# fixed marginals for sex
tally(~ sex, data=Small)
tally(~ sex, data=resample(Small, groups=sex))
# shuffled can be used to reshuffle some variables within groups
# orig.id shows where the values were in original data frame.
Small <- mutate(Small,
   id1 = paste(sex,1:10, sep=":"),
   id2 = paste(sex,1:10, sep=":"))
resample(Small, groups=sex, shuffled=c("id1","id2"))
}

#>          name birthmonth birthyear length width sex biggerfoot domhand orig.id
#> 10      Peter          4        88   24.7   8.6   B          R       L 10.10.1
#> 10.1    Peter          4        88   24.7   8.6   B          R       L  10.1.1
#> 1         Cam          3        88   27.0   9.8   B          L       R  1.1.10
#> 1.1       Cam          3        88   27.0   9.8   B          L       R 1.10.10
#> 3     Caitlin          7        88   22.5   8.6   G          R       R   3.3.5
#> 3.1   Caitlin          7        88   22.5   8.6   G          R       R   3.5.9
#> 5      Hannah          3        88   22.9   8.5   G          L       R   5.9.8
#> 9         Lee          6        88   26.7   9.0   G          L       L   9.3.3
#> 9.1       Lee          6        88   26.7   9.0   G          L       L   9.9.3
#> 8    Caroline         12        87   24.0   8.7   G          R       L   8.8.9
#>       id1  id2
#> 10   B:10  B:1
#> 10.1  B:1  B:1
#> 1     B:1 B:10
#> 1.1  B:10 B:10
#> 3     G:3  G:5
#> 3.1   G:5  G:9
#> 5     G:9  G:8
#> 9     G:3  G:3
#> 9.1   G:9  G:3
#> 8     G:8  G:9

deal(Cards, 13)    # A Bridge hand

#>  [1] "KD"  "10D" "5H"  "2H"  "7C"  "7S"  "8D"  "7D"  "QH"  "JC"  "2S"  "8C" 
#> [13] "3D" 

shuffle(Cards)

#>  [1] "7H"  "7D"  "9H"  "4S"  "QC"  "QH"  "10D" "8S"  "4H"  "10C" "9C"  "10H"
#> [13] "7C"  "KC"  "4C"  "5S"  "AS"  "9D"  "8H"  "AD"  "4D"  "7S"  "6C"  "KH" 
#> [25] "5H"  "KD"  "JH"  "AC"  "JD"  "2H"  "6S"  "QD"  "AH"  "8C"  "8D"  "KS" 
#> [37] "QS"  "6H"  "2S"  "5D"  "6D"  "JS"  "2D"  "9S"  "JC"  "3C"  "3H"  "5C" 
#> [49] "10S" "2C"  "3S"  "3D"