More Random Samples

These functions simplify and unify sampling in various ways.

resample(..., replace = TRUE)

deal(...)

shuffle(x, replace = FALSE, prob = NULL, groups = NULL,
  orig.ids = FALSE)

sample(x, size, replace = FALSE, ...)

# S3 method for default
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = FALSE, ...)

# S3 method for data.frame
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = TRUE, fixed = names(x), shuffled = c(),
  invisibly.return = NULL, ...)

# S3 method for matrix
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = FALSE, ...)

# S3 method for factor
sample(x, size, replace = FALSE, prob = NULL,
  groups = NULL, orig.ids = FALSE, drop.unused.levels = FALSE, ...)

Arguments

…
additional arguments passed to base::sample() or mosaic::sample().

replace
Should sampling be with replacement?

x
Either a vector of one or more elements from which to choose, or a positive integer.

prob
A vector of probability weights for obtaining the elements of the vector being sampled.

groups
a vector (or variable in a data frame) specifying groups to sample within. This will be recycled if necessary.

orig.ids
a logical; should original ids be included in returned data frame?

size
a non-negative integer giving the number of items to choose.

fixed
a vector of column names. These variables are shuffled en masse, preserving associations among these columns.

shuffled
a vector of column names. these variables are reshuffled individually (within groups if groups is specified), breaking associations among these columns. examples.

invisibly.return
a logical, should return be invisible?

drop.unused.levels
a logical, should unused levels be dropped?

Details

These functions are wrappers around sample() providing different defaults and natural names.

Examples

# 100 Bernoulli trials -- no need for replace=TRUE
resample(0:1, 100)

#>   [1] 1 1 1 1 0 0 1 0 0 0 0 1 0 1 1 0 0 0 0 0 1 0 0 0 0 1 1 0 0 1 1 0 0 1 0 0 0
#>  [38] 1 0 0 0 1 1 1 0 1 0 0 1 0 1 1 1 1 0 0 1 1 1 1 1 1 0 1 1 1 1 0 0 1 0 1 0 1
#>  [75] 1 1 1 0 0 0 0 1 0 0 1 0 0 1 1 1 1 1 1 0 1 1 1 1 1 0

tally(resample(0:1, 100))

#> X
#>  0  1 
#> 51 49 

if (require(mosaicData)) {
Small <- sample(KidsFeet, 10)
resample(Small)
tally(~ sex, data=resample(Small))
tally(~ sex, data=resample(Small))
# fixed marginals for sex
tally(~ sex, data=Small)
tally(~ sex, data=resample(Small, groups=sex))
# shuffled can be used to reshuffle some variables within groups
# orig.id shows where the values were in original data frame.
Small <- mutate(Small,
   id1 = paste(sex,1:10, sep=":"),
   id2 = paste(sex,1:10, sep=":"))
resample(Small, groups=sex, shuffled=c("id1","id2"))
}

#>         name birthmonth birthyear length width sex biggerfoot domhand orig.id
#> 10    Scotty          3        88   25.7   9.7   B          R       R 10.10.1
#> 1       Andy          6        88   24.0   9.2   B          R       R   1.1.1
#> 1.1     Andy          6        88   24.0   9.2   B          R       R  1.1.10
#> 1.2     Andy          6        88   24.0   9.2   B          R       R   1.1.1
#> 4   Teshanna          3        88   26.0   9.0   G          L       R   4.9.4
#> 5      Julie         11        87   26.0   9.3   G          L       R   5.9.9
#> 9     Hannah          3        88   22.9   8.5   G          L       R   9.4.4
#> 9.1   Hannah          3        88   22.9   8.5   G          L       R   9.5.5
#> 8    Heather          3        88   25.5   9.5   G          R       R   8.4.9
#> 4.1 Teshanna          3        88   26.0   9.0   G          L       R   4.8.8
#>      id1  id2
#> 10  B:10  B:1
#> 1    B:1  B:1
#> 1.1  B:1 B:10
#> 1.2  B:1  B:1
#> 4    G:9  G:4
#> 5    G:9  G:9
#> 9    G:4  G:4
#> 9.1  G:5  G:5
#> 8    G:4  G:9
#> 4.1  G:8  G:8

deal(Cards, 13)    # A Bridge hand

#>  [1] "10D" "6H"  "2H"  "8C"  "8S"  "9D"  "8D"  "KH"  "JC"  "3S"  "JS"  "3D" 
#> [13] "KD" 

shuffle(Cards)

#>  [1] "7D"  "9H"  "5S"  "QC"  "KH"  "10D" "JD"  "4H"  "10C" "9C"  "JH"  "7C" 
#> [13] "KC"  "4C"  "JC"  "8H"  "9D"  "KS"  "2H"  "4D"  "8S"  "3S"  "AH"  "6C" 
#> [25] "AD"  "5D"  "AC"  "QD"  "2D"  "5C"  "KD"  "2S"  "8C"  "3H"  "8D"  "10H"
#> [37] "5H"  "6D"  "QS"  "AS"  "6S"  "3D"  "10S" "JS"  "3C"  "6H"  "7S"  "7H" 
#> [49] "9S"  "QH"  "2C"  "4S"