Tossing Coins

These functions simplify simulating coin tosses for those (students primarily) who are not yet familiar with the binomial distributions or just like this syntax and verbosity better.

rflip(n = 1, prob = 0.5, quiet = FALSE, verbose = !quiet,
  summarize = FALSE, summarise = summarize)

# S3 method for cointoss
print(x, ...)

nflip(n = 1, prob = 0.5, ...)

Arguments

n
the number of coins to toss

prob
probability of heads on each toss

quiet
a logical. If TRUE, less verbose output is used.

verbose
a logical. If TRUE, more verbose output is used.

summarize
if TRUE, return a summary (as a data frame).

summarise
alternative spelling for summarize.

x
an object

…
additional arguments

Value

for rflip, a cointoss object

for nflip, a numeric vector

Examples

rflip(10)

#> 
#> Flipping 10 coins [ Prob(Heads) = 0.5 ] ...
#> 
#> H T T H T T T T T T
#> 
#> Number of Heads: 2 [Proportion Heads: 0.2]
#> 

rflip(10, prob = 1/6, quiet = TRUE)
rflip(10, prob = 1/6, summarize = TRUE)

#>    n heads tails      prob
#> 1 10     2     8 0.1666667

do(5) * rflip(10)

#>    n heads tails prop
#> 1 10     3     7  0.3
#> 2 10     5     5  0.5
#> 3 10     6     4  0.6
#> 4 10     5     5  0.5
#> 5 10     5     5  0.5

as.numeric(rflip(10))

#> [1] 5

nflip(10)

#> [1] 5